fix bug

+++ lib/pages/DetailPage.dart
@@ -1,4 +1,5 @@
 import 'package:akad/controller/c_Dokumen.dart';
+import 'package:akad/core_utilitas/dialog.dart';
 import 'package:akad/models/dokumen.dart';
 import 'package:akad/view_model/vm_detail.dart';
 import 'package:flutter/material.dart';
@@ -21,8 +22,8 @@ class _DetailDokumenState extends State<DetailDokumen> {
   @override
   Widget build(BuildContext context) {
     return TemplateWidget(
-        child: Scaffold(
-      body: ViewModelBuilder<DokumenVM>.reactive(
+      child: Scaffold(
+        body: ViewModelBuilder<DokumenVM>.reactive(
           viewModelBuilder: () => DokumenVM(),
           onModelReady: (vm) => vm.init(Get.parameters['id'].toString()),
           builder: (context, model, child) => !model.busy(model.data)
@@ -43,15 +44,21 @@ class _DetailDokumenState extends State<DetailDokumen> {
                             children: [
                               Padding(
                                 padding: EdgeInsets.symmetric(
-                                    vertical: 20, horizontal: 20),
-                                child: Text("Nomor Registrasi",
-                                    style: TextStyle(fontSize: 20)),
+                                  vertical: 20,
+                                  horizontal: 20,
+                                ),
+                                child: Text(
+                                  "Nomor Registrasi",
+                                  style: TextStyle(fontSize: 20),
+                                ),
                               ),
                               Padding(
                                 padding: EdgeInsets.symmetric(
                                     vertical: 20, horizontal: 50),
-                                child: Text(model.data?.dokumen.noreg ?? "",
-                                    style: TextStyle(fontSize: 18)),
+                                child: Text(
+                                  model.data?.dokumen.noreg ?? "",
+                                  style: TextStyle(fontSize: 18),
+                                ),
                               ),
                             ],
                           ),
@@ -59,17 +66,23 @@ class _DetailDokumenState extends State<DetailDokumen> {
                             children: [
                               Padding(
                                 padding: EdgeInsets.symmetric(
-                                    vertical: 20, horizontal: 20),
-                                child: Text("Keterangan Dokumen",
-                                    style: TextStyle(fontSize: 20)),
+                                  vertical: 20,
+                                  horizontal: 20,
+                                ),
+                                child: Text(
+                                  "Keterangan Dokumen",
+                                  style: TextStyle(fontSize: 20),
+                                ),
                               ),
                               Padding(
                                 padding: EdgeInsets.symmetric(
                                     vertical: 20, horizontal: 15),
                                 child: (model.data?.dokumen.keterangan == '')
                                     ? Text("-")
-                                    : Text(model.data?.dokumen.keterangan ?? "",
-                                        style: TextStyle(fontSize: 20)),
+                                    : Text(
+                                        model.data?.dokumen.keterangan ?? "",
+                                        style: TextStyle(fontSize: 20),
+                                      ),
                               ),
                             ],
                           ),
@@ -77,16 +90,24 @@ class _DetailDokumenState extends State<DetailDokumen> {
                             children: [
                               Padding(
                                 padding: EdgeInsets.symmetric(
-                                    vertical: 20, horizontal: 20),
-                                child: Text("Jenis Dokumen",
-                                    textAlign: TextAlign.left,
-                                    style: TextStyle(fontSize: 20)),
+                                  vertical: 20,
+                                  horizontal: 20,
+                                ),
+                                child: Text(
+                                  "Jenis Dokumen",
+                                  textAlign: TextAlign.left,
+                                  style: TextStyle(fontSize: 20),
+                                ),
                               ),
                               Padding(
                                 padding: EdgeInsets.symmetric(
-                                    vertical: 20, horizontal: 70),
-                                child: Text(model.data?.singkatan ?? "",
-                                    style: TextStyle(fontSize: 18)),
+                                  vertical: 20,
+                                  horizontal: 70,
+                                ),
+                                child: Text(
+                                  model.data?.singkatan ?? "",
+                                  style: TextStyle(fontSize: 18),
+                                ),
                               ),
                             ],
                           ),
@@ -94,42 +115,53 @@ class _DetailDokumenState extends State<DetailDokumen> {
                             children: [
                               Padding(
                                 padding: EdgeInsets.symmetric(
-                                    vertical: 20, horizontal: 20),
-                                child: Text("File Dokumen",
-                                    textAlign: TextAlign.left,
-                                    style: TextStyle(fontSize: 20)),
+                                  vertical: 20,
+                                  horizontal: 20,
+                                ),
+                                child: Text(
+                                  "File Dokumen",
+                                  textAlign: TextAlign.left,
+                                  style: TextStyle(fontSize: 20),
+                                ),
                               ),
                               Padding(
                                 padding: EdgeInsets.symmetric(
                                     vertical: 20, horizontal: 82.5),
-                                child: Text(model.data?.dokumen.nama ?? "",
-                                    style: TextStyle(fontSize: 18)),
+                                child: Text(
+                                  model.data?.dokumen.nama ?? "",
+                                  style: TextStyle(fontSize: 18),
+                                ),
                               ),
                             ],
                           ),
-                          if(model.data != null)
-                          Container(
-                            margin: EdgeInsets.only(left: 82.5),
-                            child: (model.getExtension(
-                                        model.data?.dokumen.nama ?? "") ==
-                                    'pdf')
-                                ? Container(
-                                    width: 650,
-                                    height: 400,
-                                    margin: const EdgeInsetsDirectional.all(25),
-                                    child: PdfPreview(
-                                      build: (newFile) => generatePDF(newFile,
-                                          model.getFile(model.data!.dokumen)),
-                                      
-                                    ),)
-                                : Container(
-                                    width: 650,
-                                    height: 350,
-                                    margin: const EdgeInsetsDirectional.all(25),
-                                    child: Image.memory(
-                                        model.getFile(model.data!.dokumen)),
-                                  ),
-                          ),
+                          if (model.data != null)
+                            Container(
+                              margin: EdgeInsets.only(left: 82.5),
+                              child: (model.getExtension(
+                                          model.data?.dokumen.nama ?? "") ==
+                                      'pdf')
+                                  ? Container(
+                                      width: 650,
+                                      height: 400,
+                                      margin:
+                                          const EdgeInsetsDirectional.all(25),
+                                      child: PdfPreview(
+                                        build: (newFile) => generatePDF(
+                                          newFile,
+                                          model.getFile(model.data!.dokumen),
+                                        ),
+                                      ),
+                                    )
+                                  : Container(
+                                      width: 650,
+                                      height: 350,
+                                      margin:
+                                          const EdgeInsetsDirectional.all(25),
+                                      child: Image.memory(
+                                        model.getFile(model.data!.dokumen),
+                                      ),
+                                    ),
+                            ),
                           Container(
                             margin: EdgeInsets.only(left: 107.5),
                             width: 650,
@@ -142,7 +174,8 @@ class _DetailDokumenState extends State<DetailDokumen> {
                                 ElevatedButton(
                                   onPressed: () => {
                                     Get.toNamed(
-                                        '/edit/${Get.parameters['id'].toString()}')
+                                      '/edit/${Get.parameters['id'].toString()}',
+                                    )
                                   },
                                   child: Text('Edit'),
                                 ),
@@ -151,10 +184,20 @@ class _DetailDokumenState extends State<DetailDokumen> {
                                   style: ElevatedButton.styleFrom(
                                     primary: Colors.red,
                                   ),
-                                  onPressed: () {
-                                    deleteDokumen(
-                                        Get.parameters['id'].toString());
-                                    Get.offAllNamed('/home');
+                                  onPressed: () async {
+                                    bool confirm = await showConfirmDialog(
+                                      subtitle: "Anda yakin akan menghapus ?",
+                                      styleTextYes: TextStyle(
+                                        color: Colors.red,
+                                      ),
+                                    );
+
+                                    if (confirm) {
+                                      deleteDokumen(
+                                        Get.parameters['id'].toString(),
+                                      );
+                                      Get.offAllNamed('/home');
+                                    }
                                   },
                                   child: Text('Hapus'),
                                 ),
@@ -166,7 +209,9 @@ class _DetailDokumenState extends State<DetailDokumen> {
                     ),
                   ),
                 )
-              : Text("loading...")),
-    ));
+              : Text("loading..."),
+        ),
+      ),
+    );
   }
 }
